Add files via upload
Solving problems after refreshing credentials triggering a new mqtt connection with the new credentials. Because the new connection was triggered in the onMqttDisconnect(), the mqtt loop was not started. Reason: race condition! onMqttDisconnect is called by mqtt loop at the end of disconnection process... While the mqtt loop thread was not yet stopped, it happened that the loop was already started. Which did not work of course...